Extract plugin build tests to not depend on composite build testing base
Because those tests use different fixtures that get in the way when including plugin builds with custom configuration
