+++ subprojects/composite-builds/src/integTest/groovy/org/gradle/integtests/composite/AbstractCompositeBuildIntegrationTest.groovy
@@ -205,57 +205,4 @@ public class ${className} implements Plugin<Project> {
         def output = result.output
         assert output.contains(string.trim())
     }
-
-    PluginBuildFixture pluginBuild(String buildName) {
-        return new PluginBuildFixture(buildName)
-    }
-
-    class PluginBuildFixture {
-        final String buildName
-        final String settingsPluginId
-        final String projectPluginId
-
-        final TestFile settingsFile
-        final TestFile buildFile
-
-        PluginBuildFixture(String buildName) {
-            this.buildName = buildName
-            this.settingsPluginId = "${buildName}.settings-plugin"
-            this.projectPluginId = "${buildName}.project-plugin"
-            this.settingsFile = file("$buildName/settings.gradle")
-            this.buildFile = file("$buildName/build.gradle")
-
-            settingsFile << """
-                rootProject.name = "$buildName"
-            """
-            buildFile << """
-                plugins {
-                    id("groovy-gradle-plugin")
-                }
-            """
-            file("$buildName/src/main/groovy/${projectPluginId}.gradle") << """
-                println('$projectPluginId applied')
-            """
-            file("$buildName/src/main/groovy/${settingsPluginId}.settings.gradle") << """
-                println('$settingsPluginId applied')
-            """
-        }
-
-        void assertSettingsPluginApplied() {
-            outputContains("$settingsPluginId applied")
-        }
-
-        void assertSettingsPluginNotApplied() {
-            outputDoesNotContain("$settingsPluginId applied")
-        }
-
-        void assertProjectPluginApplied() {
-            outputContains("$projectPluginId applied")
-        }
-
-        void assertProjectPluginNotApplied() {
-            outputDoesNotContain("$projectPluginId applied")
-        }
-    }
-
 }
