+++ subprojects/composite-builds/src/integTest/groovy/org/gradle/integtests/composite/plugins/PluginBuildsOrderingIntegrationTest.groovy
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.integtests.composite
+package org.gradle.integtests.composite.plugins
 
-class CompositePluginBuildsOrderingIntegrationTest extends AbstractCompositeBuildIntegrationTest {
+class PluginBuildsOrderingIntegrationTest extends AbstractPluginBuildIntegrationTest {
 
     private final PluginBuildFixture build1 = pluginBuild("logic-1")
     private final PluginBuildFixture build2 = pluginBuild("logic-2")
